# api/openapi.json
{
  "components": {
    "parameters": {
      "Consistency": {
        "in": "query",
        "name": "consistency",
        "schema": {
          "default": "require_fresh",
          "enum": [
            "require_fresh",
            "allow_stale",
            "cache_only"
          ],
          "type": "string"
        }
      },
      "Refresh": {
        "in": "query",
        "name": "refresh",
        "schema": {
          "default": false,
          "type": "boolean"
        }
      }
    },
    "responses": {
      "Error": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "Stable API failure"
      }
    },
    "schemas": {
      "APIError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Quote": {
        "properties": {
          "Amount": {
            "type": "number"
          },
          "Code": {
            "type": "string"
          },
          "High": {
            "type": "number"
          },
          "LastClose": {
            "type": "number"
          },
          "Low": {
            "type": "number"
          },
          "Name": {
            "type": "string"
          },
          "Open": {
            "type": "number"
          },
          "Price": {
            "type": "number"
          },
          "Volume": {
            "type": "number"
          }
        },
        "required": [
          "Code",
          "Price"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Canonical transport contract. Legacy endpoints retain their existing success payload while all failures use ErrorEnvelope.",
    "title": "TongStock API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/quote": {
      "get": {
        "operationId": "getQuote",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Consistency"
          },
          {
            "$ref": "#/components/parameters/Refresh"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            },
            "description": "Quote from the database read model"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}